requirements.txt
prettytable==0.7.2
requests==2.12.4
beautifulsoup4==4.5.3
strawpy==1.1
lxml==3.8.0
psutil==5.2.1
PythonGists==1.1.3
GitPython==2.1.3
Pillow==4.2.1
pyfiglet==0.7.5
tokage==0.1.6.5
pytz==2017.2
PyGithub==1.34
discord
slackclient
pymongo[gssapi,srv,tls]
dnspython
termcolor
GroupyAPI
